+++ app/models/notification.rb
@@ -11,6 +11,7 @@ class Notification < ApplicationRecord
   belongs_to :project
   belongs_to :adverse_event
   belongs_to :comment
+  belongs_to :export
   belongs_to :handoff
   belongs_to :sheet_unlock_request