+++ db/migrate/20161021145211_add_export_id_to_notifications.rb
@@ -1,0 +1,6 @@
+class AddExportIdToNotifications < ActiveRecord::Migration[5.0]
+  def change
+    add_column :notifications, :export_id, :integer
+    add_index :notifications, :export_id
+  end
+end